Fix major warnings
- use `qFuzzyCompare` to compare floats
- explicitly cast types when narrowing occurs
- use `Q_NULLPTR` in API and `nullptr` in code in place of `0` pointers
- add `Q_DECL_OVERRIDE` to subclasses destructors and overidden functions

+++ src/items/qniteline.cpp
@@ -26,7 +26,8 @@ bool QniteLine::select(QPoint p) {
   // get the distance from the first point on the path
   int dataSize = xMapped().size();
   for (int i = 0; i < dataSize; ++i) {
-    QPoint cp(xMapped().at(i), yMapped().at(i));
+    QPoint cp(static_cast<int>(xMapped().at(i)),
+              static_cast<int>(yMapped().at(i)));
     QPoint d = p - cp;
     if (d.manhattanLength() < SELECTION_TOLERANCE) {
       m_selected = true;
